+++ .github/workflows/tests-dev.yml
@@ -7,10 +7,21 @@ on:
 
 jobs:
   tests-dev:
+    name: ${{ matrix.suite }} (py${{ matrix.python-version }})
     runs-on: ubuntu-latest
+
     strategy:
+      fail-fast: false
       matrix:
-        python-version: ["3.11", "3.13"]
+        include:
+          - {suite: ase-tests-dev, python-version: "3.11"}
+          - {suite: ase-tests-dev, python-version: "3.13"}
+          - {suite: core-tests-dev, python-version: "3.13"}
+          - {suite: ase-dos-tests-dev, python-version: "3.13"}
+          - {suite: ase-explore-tests-dev, python-version: "3.13"}
+          # `nvalchemi-tests` is left out on purpose: combining the metatrain
+          # and nvalchemi-toolkit development versions makes failures hard to
+          # attribute to either one
 
     steps:
     - uses: actions/checkout@v6
@@ -26,8 +37,8 @@ jobs:
       run: python -m pip install tox
 
     - name: run Python tests
-      run: |
-        tox -e upet-tests-dev
+      run: tox -e ${{ matrix.suite }}
       env:
         # Use the CPU only version of torch when building/running the code
         PIP_EXTRA_INDEX_URL: https://download.pytorch.org/whl/cpu
+        HF_TOKEN: ${{ secrets.HF_TOKEN }}
