+++ AGENTS.md
@@ -8,16 +8,20 @@ All workflows are driven through `tox` (see `tox.ini`). Install it once with `pi
 
 - `tox -e lint` — ruff format check, ruff lint, mypy, and sphinx-lint on `src/` and `tests/` plus `README.md`.
 - `tox -e format` — auto-apply `ruff format` and `ruff check --fix-only` to the same paths.
-- `tox -e upet-tests` — run the main package tests (`tests/upet/`) against the pinned `metatrain` release.
-- `tox -e upet-tests-dev` — same suite but installs `metatrain` from git `main` first; used by the weekly CI job.
-- `tox -e pet-mad-dos-tests` — run the PET-MAD-DOS tests (`tests/pet_mad_dos/`).
+- `tox -e core-tests` — backend-agnostic tests (`tests/core/`): model registry, metadata, deprecated shims.
+- `tox -e ase-tests` — ASE calculator tests (`tests/ase/`, excluding `dos/` and `explore/`).
+- `tox -e ase-dos-tests` — PET-MAD-DOS tests (`tests/ase/dos/`).
+- `tox -e ase-explore-tests` — PET-MAD explorer tests (`tests/ase/explore/`).
+- `tox -e nvalchemi-tests` — NVAlchemi wrapper tests (`tests/nvalchemi/`); installs the optional `nvalchemi` extra, without which the whole suite `importorskip`s itself.
+  On macOS the `test_wrapper_compile.py` tests cannot run: they make inductor build a C++ kernel that loads a second `libomp.dylib`, and OpenMP aborts the whole interpreter (`OMP: Error #15`) rather than failing a single test. Setting `KMP_DUPLICATE_LIB_OK=TRUE` only trades the abort for a deadlock. Deselect them locally (`tox -e nvalchemi-tests -- -k 'not CompiledBackend'`) and rely on the Linux CI job for that coverage.
+- `tox -e {core,ase,ase-dos,ase-explore}-tests-dev` — the same suites, but installing `metatrain` from git `main` first; used by the weekly CI job.
 - `tox -e build` — build the sdist + wheel and run `twine check` and `check-manifest`.
 - `tox -e docs` — build the Sphinx HTML docs into `docs/build/html` (runs with `--fail-on-warning`).
-- Single test: `tox -e upet-tests -- -k test_name` (everything after `--` is forwarded as `{posargs}` to pytest; `changedir` is already `tests/upet`, so relative paths resolve there).
+- Single test: `tox -e ase-tests -- -k test_name` (everything after `--` is forwarded as `{posargs}` to pytest; each suite has its own `changedir`, so relative paths resolve inside that suite's directory).
 
 `pytest` is configured with `filterwarnings = ["error", ...]` in `pyproject.toml`, so any new warning that isn't explicitly ignored will fail tests — add entries to that allowlist rather than silencing warnings inline.
 
-CI (`.github/workflows/`) runs `tox -e upet-tests` and `tox -e pet-mad-dos-tests` on push/PR across Linux/macOS/Windows + Python 3.11/3.13 with `PIP_EXTRA_INDEX_URL=https://download.pytorch.org/whl/cpu`. Use the same env var locally if you hit CUDA-wheel install issues. `HF_TOKEN` is needed to exercise tests that pull gated checkpoints.
+CI (`.github/workflows/tests.yml`) runs one job per suite, named `<suite> (<os>, py<version>)`, on push/PR. `ase-tests` covers the full matrix (Linux 3.11/3.13, macOS 3.13, Windows 3.13); the other suites run once on Linux + 3.13. All of them set `PIP_EXTRA_INDEX_URL=https://download.pytorch.org/whl/cpu` — use the same env var locally if you hit CUDA-wheel install issues — and `HF_TOKEN`, which is needed to exercise tests that pull gated checkpoints. `tests-dev.yml` runs the `-dev` variants weekly.
 
 ## Architecture
 
@@ -39,11 +43,16 @@ Package layout (`src/upet/`):
 
 Tests (`tests/`):
 
-- `tests/upet/` covers the calculator, MD, non-conservative forces, uncertainty, rotational averaging, featurizer, offline checkpoints, and metadata — parametrized over `UPET_AVAILABLE_MODELS`.
-- `tests/pet_mad_dos/` covers DOS-specific paths.
+`tests/` mirrors `src/upet/`, one directory (and one tox environment) per source sub-package:
+
+- `tests/core/` — ecosystem-agnostic code: model registry (`test_models.py`), checkpoint-name parsing (`test_checkpoint_names.py`), metadata (`test_metadata.py`), and the deprecated top-level shims (`test_deprecated_imports.py`).
+- `tests/ase/` — `UPETCalculator`: energies/forces/stress, MD, non-conservative forces, uncertainty, rotational averaging — parametrized over `UPET_AVAILABLE_MODELS`. `_utils.py` holds the shared non-conservative-support expectations, imported as `from _utils import ...` (which works because pytest puts each test file's own directory on `sys.path`).
+- `tests/ase/dos/` and `tests/ase/explore/` — the DOS and explorer byproducts.
+- `tests/nvalchemi/` — `UPETWrapper`, split by concern: `test_wrapper_construction.py` (`__init__`, buffers, `ModelConfig`, properties), `test_wrapper_adapters.py` (`adapt_input` / `adapt_output`), `test_wrapper_forward.py` (`forward`, `compute_embeddings`), `test_wrapper_compile.py` (`torch.compile` parity), `test_wrapper_export.py` (`export_model`) and `test_wrapper_checkpoint.py` (integration against a real HuggingFace checkpoint). Shared constants, hypers and `AtomicData` builders live in `_helpers.py`; the fixtures in `conftest.py`. Every `test_wrapper_*.py` opens with `pytest.importorskip("nvalchemi")`, so the whole directory skips without the extra — `conftest.py` therefore imports `nvalchemi` only inside fixture bodies, never at module scope.
+- Test file basenames are unique across the whole tree (hence `test_dos_metadata.py`, not a second `test_metadata.py`): there are no `__init__.py` files, so pytest imports test modules into a flat namespace and duplicate basenames collide when running `pytest tests/` from the repository root.
 - Each suite has its own `changedir` in `tox.ini`; tests pull real checkpoints from HuggingFace, so they are network-bound by design.
 
-External dependencies to keep in mind: `metatrain` (pinned to `>=2026.2,<2026.3`), `metatomic-ase`, `nvalchemi-toolkit-ops` (pinned exactly to `0.2.0`), and `huggingface_hub`. Version bumps to these usually require matching updates to the warning allowlist in `pyproject.toml` and sometimes to `_version.py`.
+External dependencies to keep in mind: `metatrain` (pinned to `>=2026.4,<2026.5`), `metatomic-ase`, `nvalchemi-toolkit-ops` (pinned to `>=0.4.0,<0.5.0`), and `huggingface_hub`. Version bumps to these usually require matching updates to the warning allowlist in `pyproject.toml` and sometimes to `_version.py`.
 
 ## Documentation
