+++ pyproject.toml
@@ -135,6 +135,13 @@ filterwarnings = [
     "ignore:`compute_requested_neighbors` is deprecated.*:UserWarning",
     "ignore:`compute_requested_neighbors_from_options` is deprecated.*:UserWarning",
     "ignore: found nvalchemi-toolkit-ops version 0.4.0, this code was only tested with version >=0.3.0,<=0.4.0.*:UserWarning",
+    # warp, imported by nvalchemi before `upet` gets a chance to filter this
+    # itself (see the same filter in `src/upet/__init__.py`)
+    "ignore:warp.config.quiet is deprecated:DeprecationWarning",
+    # nvalchemi `AtomicData` casts every float field to the dtype of
+    # `positions`; harmless when a test builds float64 data from float32
+    # defaults it never sets itself (`atomic_masses`, `velocities`)
+    "ignore:AtomicData fields .* were cast from their original dtypes:UserWarning",
     # HF Xet deprecation warning
     "ignore:hf_xet.download_files\\(\\) is deprecated. Use XetSession\\(\\):DeprecationWarning",
     # Numpy warning
