+++ tests/core/test_checkpoint_names.py
