+++ tests/core/test_deprecated_imports.py
