+++ tests/core/test_metadata.py
