+++ tox.ini
@@ -4,10 +4,21 @@ requires =
     tox>=4.31 
 # these are the default environments, i.e. the list of tests running when you
 # execute `tox` in the command-line without anything else
-envlist = lint, build, upet-tests, upet-tests-dev, pet-mad-dos-tests, pet-mad-tests-dev
+envlist =
+    lint
+    build
+    core-tests
+    ase-tests
+    ase-dos-tests
+    ase-explore-tests
+    nvalchemi-tests
 
 [testenv]
 passenv = *
+# defaults shared by all the test suites below; the lint/format/build/docs
+# environments override both of these with their own
+deps = pytest
+commands = pytest {posargs}
 lint_folders =
     "{toxinidir}/docs/src/" \
     "{toxinidir}/examples/" \
@@ -39,35 +50,49 @@ commands =
     ruff format {[testenv]lint_folders}
     ruff check --fix-only {[testenv]lint_folders} "{toxinidir}/README.md" {posargs}
 
-[testenv:upet-tests]
-description = Run basic package tests with pytest
-passenv = *
-deps =
-    pytest
+# --------------------------------------------------------------------------- #
+# Test suites, one per source sub-package (see `src/upet/`)
+# --------------------------------------------------------------------------- #
 
-changedir = tests/upet
-allowlist_externals = bash
-commands =
-    pytest {posargs}
+[testenv:core-tests]
+description = Run the backend-agnostic tests (model registry, metadata, shims)
+changedir = tests/core
 
-[testenv:upet-tests-dev]
-description = Run basic package tests against metatrain dev version
-deps =
-    pytest
-changedir = tests/upet
-commands =
+[testenv:ase-tests]
+description = Run the ASE calculator tests (`upet.ase`)
+changedir = tests/ase
+# `dos` and `explore` have their own environments below
+commands = pytest --ignore=dos --ignore=explore {posargs}
+
+[testenv:ase-dos-tests]
+description = Run the PET-MAD-DOS tests (`upet.ase.dos`)
+changedir = tests/ase/dos
+
+[testenv:ase-explore-tests]
+description = Run the PET-MAD explorer tests (`upet.ase.explore`)
+changedir = tests/ase/explore
+
+[testenv:nvalchemi-tests]
+description = Run the NVAlchemi wrapper tests (`upet.nvalchemi`)
+extras = nvalchemi
+changedir = tests/nvalchemi
+
+[testenv:{core,ase,ase-dos,ase-explore}-tests-dev]
+description = Run the matching test suite against the metatrain dev version
+changedir =
+    core: tests/core
+    dos: tests/ase/dos
+    explore: tests/ase/explore
+    ase-!dos-!explore: tests/ase
+commands_pre =
     pip install "metatrain @ git+https://github.com/metatensor/metatrain.git"
-    pytest {posargs}
-
-[testenv:pet-mad-dos-tests]
-description = Run basic package tests with pytest
-deps =
-    pytest
-
-changedir = tests/pet_mad_dos
-allowlist_externals = bash
 commands =
-    pytest {posargs}
+    ase-!dos-!explore: pytest --ignore=dos --ignore=explore {posargs}
+    !ase: pytest {posargs}
+    dos: pytest {posargs}
+    explore: pytest {posargs}
+
+# --------------------------------------------------------------------------- #
 
 [testenv:build]
 description = builds the package and checks integrity
